Disk NRAW backend incorporated into picviz * extractor does not work * Pure mapping filters are defined, as they will be (in a close futur) parallelised in the TBB normalisation pipeline * Adapt mappign filters (lots of room for improvements here) * pvconfig.ini: nraw_tmp for nraw temporary directory

+++ libpicviz/src/include/picviz/PVSelection.h
@@ -402,7 +402,7 @@ public:
 			sse_sel = _mm_load_si128((__m128i*) &_table[i]);
 			PVCore::PVBitVisitor::visit_bits(sse_sel, f, chunk_to_line_index(i));
 		}
-		for (; i < last_chunk; i++) {
+		for (; i < (size_t) last_chunk; i++) {
 			const uint64_t sel_buf = _table[i];
 			PVCore::PVBitVisitor::visit_bits(sel_buf, f, chunk_to_line_index(i));
 		}
